test/requirements.txt
webdriver-manager==4.0.2
selenium==4.36.0
pytest==8.4.2
pytest-playwright==0.7.1